+++ utilities/interp/interp_deps.m
@@ -1,30 +1,31 @@
-function varDeps = interp_deps(Mobj, varRec, depRec)
-% 
+function varDeps = interp_deps(depRec, varRec, depLayers)
+% interpolate data from standard layers onto sigma layers
 % 
 %% Syntax
-% 
+% varDeps = interp_deps(Mobj, varRec, depRec)
 %
 %% Description 
-% 
-%
-%% Examples
-%
+% varDeps = interp_deps(Mobj, varRec, depRec)
 %
 %% Input Arguments
-%
+% varRec - original variable matrix; double
+%       variable matrix (nNodes*nz) on standard layers;
+% depRec - standard depths; double
+%       standard depth layers;
+% depLayers - depth matrix; double
+%       the depth matrix (maxLev*nNodes), indicating depth layers for all
+%       nodes. It is typically from Mobj.
 %
 %% Output Arguments
-% 
-% 
-%% Notes
-%
+% varDeps - interpolated variable matrix; double
+%       varaible matrix (nNodes*maxLev) on sigma layers
 %
 %% Author Info
-% Created by Wenfan Wu, Ocean Univ. of China in 2023. 
-% Last Updated on 2023-11-26.
-% Email: wenfanwu@stu.ouc.edu.cn
-% 
-% See also: 
+% Created by Wenfan Wu, Virginia Institute of Marine Science in 2023.
+% Last Updated on 09 Dec 2024.
+% Email: wwu@vims.edu
+%
+% See also: interp_tri
 
 %% Parse inputs
 if size(varRec, 2) ~= length(depRec)
@@ -33,14 +34,16 @@ end
 depRec = abs(depRec);
 depRec = depRec-min(depRec);
 
-depLayers = abs(Mobj.depLayers);
-varDeps = nan(Mobj.nNodes, Mobj.maxLev);
-for iNode = 1:Mobj.nNodes
-    progressbar(iNode/Mobj.nNodes)
+depLayers = abs(depLayers);
+[maxLev, nNodes] = size(depLayers);
+
+varDeps = nan(nNodes, maxLev);
+for iNode = 1:nNodes
+    progressbar(iNode/nNodes)
     
     depTri = depLayers(:, iNode);
     varTmp = varRec(iNode,:);
-    varDeps(iNode,:) = multi_interp1(depRec, varTmp, depTri, 1);
+    varDeps(iNode,:) = interp1(depRec, varTmp, depTri);
 end
 varDeps = fillmissing(varDeps, 'previous', 2, 'EndValues', 'previous');
 end